refactor(plan-detail): extract useRedirects into shell/hooks
